parse ctor sig

+++ crates/compiler/src/ast/node_flags.rs
@@ -1,6 +1,6 @@
 bitflags::bitflags! {
   #[derive(Clone, Copy, Debug)]
-  pub struct NodeFlags: u32 {
+  pub struct NodeFlags: u8 {
     const LET   = 1 << 0;
     const CONST = 1 << 1;
     const USING = 1 << 2;
